Move go.mod to top-level folder.
